pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-ssh-orchestrator"
version = "0.8.0"
description = "A secure SSH fleet orchestrator for MCP (STDIO transport)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Sammy Farida" }]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "mcp>=1.21.0,<2.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "paramiko>=3.5.0,<5.0.0",
]

[project.optional-dependencies]
test = [
  "pytest>=7.0.0,<10.0.0",
  "pytest-cov>=4.0.0,<8.0.0",
  "pytest-mock>=3.10.0,<4.0.0",
]
dev = [
  "ruff>=0.1.0,<1.0.0",
  "black>=24.0.0,<26.0.0",
  "mypy>=1.0.0,<2.0.0",
  "pip-tools>=7.0.0,<8.0.0",
]

[project.scripts]
mcp-ssh-orchestrator = "mcp_ssh.mcp_server:main"

[tool.setuptools.package-data]
mcp_ssh = ["py.typed"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
]
ignore = [
  "E501",  # line too long, handled by black
  "B008",  # do not perform function calls in argument defaults
  "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py311']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--strict-markers",
  "--strict-config",
  "--verbose",
]
